requirements.txt
flask
gunicorn
openai
langchain-openai
langchain-pinecone
flask
python-dotenv
gunicorn
wfastcgi